+++ issuer-application/src/test/java/ch/admin/bj/swiyu/issuer/oid4vci/intrastructure/web/controller/DeferredFlowIT.java
@@ -385,27 +385,6 @@ class DeferredFlowIT {
     }
 
     @Test
-    void testOfferCreation_withUnexpectedClaim() throws Exception {
-
-        var extendedOfferData = new HashMap<String, Object>(getUniversityCredentialSubjectData());
-        extendedOfferData.put("unexpectedClaim", "unexpectedValue");
-
-        var offerRequest = CreateCredentialOfferRequestDto.builder()
-                .metadataCredentialSupportedId(List.of("university_example_sd_jwt"))
-                .credentialMetadata(getDeferredCredentialMetadataDto())
-                .credentialSubjectData(extendedOfferData)
-                .build();
-
-        var offerRequestString = objectMapper.writeValueAsString(offerRequest);
-
-        TestInfrastructureUtils.createCredentialOffer(mock, offerRequestString)
-                .andExpect(status().is4xxClientError())
-                .andExpect(jsonPath("$.detail")
-                        .value("Unexpected additional credential claims found: [unexpectedClaim]"))
-                .andReturn();
-    }
-
-    @Test
     void testOfferCreation_withMissingMandatoryClaim() throws Exception {
 
         var extendedOfferData = new HashMap<String, Object>(getUniversityCredentialSubjectData());