+++ CHANGELOG.md
@@ -41,7 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Removed Deprecated OID4VCI Draft 13 Endpoints
     - Removed support for did:jwk, as it is not part of the swiss-profiles anymore
 - Use OID4VCI 1.0 compliant error codes for credential_endpoint and deferred_credential_endpoint error responses.
-- Validation uses now the `credential_metadata.claims` as default for the validation and the `claims` as fallback.
+- Validation uses now the `credential_metadata.claims` as default for the validation and the `claims` as fallback amd
+  don't check surplus.
 - Contracted cnf to now only provide the correct shape as defined in RFC 7800
 
 ### Fixed